+++ integration_tests/models/test_generate_dimension_type1.sql
@@ -1,0 +1,6 @@
+{{ clearpoint_dbt_utils.generate_dimension(
+    source_relation=ref('test_dedupe_latest_record'),
+    business_key='customer_id',
+    surrogate_key_alias='customer_key',
+    attribute_columns=['name', 'email']
+) }}